slightly refactoring
