+++ PowerPointLabs/PowerPointLabs/ImageSearch/ImageSearchPane.xaml.OnLoadImages.cs
@@ -2,10 +2,8 @@
 using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
-using System.Windows;
 using System.Windows.Forms;
 using PowerPointLabs.AutoUpdate;
-using PowerPointLabs.ImageSearch.Crop;
 using PowerPointLabs.ImageSearch.Domain;
 using PowerPointLabs.ImageSearch.SearchEngine;
 using PowerPointLabs.ImageSearch.SearchEngine.VO;
@@ -107,38 +105,6 @@ namespace PowerPointLabs.ImageSearch
                     ShowErrorMessageBox(TextCollection.ImagesLabText.ErrorImageCorrupted);
                 }
             }));
-        }
-
-        private void CropImage(ImageItem imageItem = null)
-        {
-            if (imageItem == null && SearchListBox.SelectedValue == null) return;
-
-            var source = imageItem ?? (ImageItem) SearchListBox.SelectedValue;
-            var cropWindow = new CropWindow();
-            cropWindow.SetThumbnailImage(source.ImageFile);
-            cropWindow.SetFullsizeImage(source.FullSizeImageFile);
-            if (source.Rect.Width > 1)
-            {
-                cropWindow.SetCropRect(source.Rect.X, source.Rect.Y, source.Rect.Width, source.Rect.Height);
-            }
-            cropWindow.ShowDialog();
-
-            if (cropWindow.IsCropped)
-            {
-                source.CroppedImageFile = cropWindow.CropResult;
-                source.CroppedThumbnailImageFile = cropWindow.CropResultThumbnail;
-                source.Rect = cropWindow.Rect;
-
-                var imageIndex = SearchListBox.Items.IndexOf(source);
-                if (imageIndex >= 0
-                    && imageIndex < _downloadedImages.Count)
-                {
-                    var imageToPersist = _downloadedImages[imageIndex];
-                    imageToPersist.CroppedImageFile = source.CroppedImageFile;
-                    imageToPersist.CroppedThumbnailImageFile = source.CroppedThumbnailImageFile;
-                    imageToPersist.Rect = source.Rect;
-                }
-            }
         }
 
         private static void VerifyIsProperImage(string filename)
@@ -367,27 +333,5 @@ namespace PowerPointLabs.ImageSearch
             return VOUtil.GetTitle(searchResult) + "\n" + VOUtil.GetWidth(searchResult) + " x " + VOUtil.GetHeight(searchResult);
         }
         # endregion
-
-        private void MenuItemCrop_OnClick(object sender, RoutedEventArgs e)
-        {
-            if (SearchButton.SelectedIndex == TextCollection.ImagesLabText.ButtonIndexSearch) return;
-
-            if (_rightClickedSearchListBoxItemIndex < 0
-                || _rightClickedSearchListBoxItemIndex > SearchListBox.Items.Count)
-                return;
-
-            var selectedImage = (ImageItem)SearchListBox.Items.GetItemAt(_rightClickedSearchListBoxItemIndex);
-            if (selectedImage == null) return;
-
-            CropImage(selectedImage);
-        }
-
-        private void MenuItemCrop_OnClickFromPreviewListBox(object sender, RoutedEventArgs e)
-        {
-            var selectedImage = (ImageItem)SearchListBox.SelectedItem;
-            if (selectedImage == null || selectedImage.ImageFile == StoragePath.LoadingImgPath) return;
-
-            CropImage(selectedImage);
-        }
     }
 }
