+++ PowerPointLabs/PowerPointLabs/ImageSearch/ImageSearchPane.xaml.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Collections.Specialized;
-using System.ComponentModel;
 using System.Reflection;
 using System.Windows;
 using System.Windows.Controls;
@@ -201,29 +200,6 @@ namespace PowerPointLabs.ImageSearch
             PreviewList = new ObservableCollection<ImageItem>();
             PreviewList.CollectionChanged += PreviewList_OnCollectionChanged;
             PreviewListBox.DataContext = this;
-        }
-
-        private void InitSearchList()
-        {
-            SearchList = new ObservableCollection<ImageItem>();
-            _downloadedImages = StoragePath.Load(ImagesLabImagesList);
-            SearchList.CollectionChanged += SearchList_OnCollectionChanged;
-
-            foreach (var imageItem in _downloadedImages)
-            {
-                _filesInUse.Add(imageItem.ImageFile);
-                _filesInUse.Add(imageItem.FullSizeImageFile);
-                if (imageItem.CroppedImageFile != null)
-                {
-                    _filesInUse.Add(imageItem.CroppedImageFile);
-                    _filesInUse.Add(imageItem.CroppedThumbnailImageFile);
-                }
-            }
-            CopyContentToObservableList(_downloadedImages, SearchList);
-            Dispatcher.BeginInvoke(new Action(() =>
-            {
-                SearchListBox.DataContext = this;
-            }));
         }
 
         private void InitSearchOptions()
@@ -458,17 +434,6 @@ namespace PowerPointLabs.ImageSearch
             }
         }
 
-        // rmb to close background presentation
-        private void ImageSearchPane_OnClosing(object sender, CancelEventArgs e)
-        {
-            IsOpen = false;
-            if (PreviewPresentation != null)
-            {
-                PreviewPresentation.Close();
-            }
-            StoragePath.Save(ImagesLabImagesList, _downloadedImages);
-        }
-
         private void AdvancedButton_OnClick(object sender, RoutedEventArgs e)
         {
             SearchOptionsFlyout.IsOpen = true;
@@ -533,24 +498,6 @@ namespace PowerPointLabs.ImageSearch
                     SearchTextboxWatermark.Text = TextCollection.ImagesLabText.TextBoxWatermarkFromFile;
                     SearchInstructions.Text = TextCollection.ImagesLabText.InstructionForFromFile;
                     break;
-            }
-        }
-
-        private void CopyContentToObservableList(IEnumerable<ImageItem> content, ObservableCollection<ImageItem> target)
-        {
-            foreach (var image in content)
-            {
-                target.Add(new ImageItem
-                {
-                    ImageFile = image.ImageFile,
-                    FullSizeImageFile = image.FullSizeImageFile,
-                    FullSizeImageUri = image.FullSizeImageUri,
-                    ContextLink = image.ContextLink,
-                    Tooltip = image.Tooltip,
-                    CroppedImageFile = image.CroppedImageFile,
-                    CroppedThumbnailImageFile = image.CroppedThumbnailImageFile,
-                    Rect = image.Rect
-                });
             }
         }
